Corrected probability of discontinuing following no event (also in generate.transition.matrix.7.R)
Changed generate.transition.matrix.7.apixref.R and generate.transition.matrix.7.R. If the probability of discontinuing following no event, TIA, and SE were x, y, and z and if the probability of having each of these events were a, b, and c, my original code was doing
(a+b+c)*(x+y+z)
which is wrong. The correction is now doing
a*x+b*y+c*z
which is correct.